Fix some incorrect fields
Because of how things were laid out instance-vs-static-wise, it was not
very simple to send things like App, Device based on the property values
on the actual IBugsnagger instance. The seemingly obvious thing to do
was to still use `Bugsnag::GetEvent`, which unfortunately pulled from
static fields on the Bugsnag class.

To fix this I added some extension methods for `IBugsnaggers` in order
to set the App/Device properties on the event correctly such cases.