Add (still failing) test for asyncawait lines
They're failing, because the "we can't parse anything at all" case isn't
quite right yet.

Thinking about it, in a stack trace, the minimum we seem to get (from
the framework line case) is a method name, without file or line. This
makes sense because we probably don't just have the framework pdbs
around when we send out the stack trace, or any reason to know details
the source of the framework assemblies.

Because of this, it seems like the "bare minimum" should just go in the
method name property... If we can't even seem to parse that, we should
just include the entire contents of the line as the method name. This
way errors reported to Bugsnag can at least tell us what came in on that
line...