Use whole line as method if we can't parse
This operates under the opinion that method name is the "bare minimum"
of the info we can get out of a stack trace.

If we can't even seem to parse something sensible for that, let's just
put the whole line in there, so that the incoming value can at least
show up on the reporting side.